[Bronze V] Title: 이상한 기호, Time: 128 ms, Memory: 14228 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -10,8 +10,8 @@ public class Main {
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int N = Integer.parseInt(st.nextToken());
-        int M = Integer.parseInt(st.nextToken());
+        long N = Long.parseLong(st.nextToken());
+        long M = Long.parseLong(st.nextToken());
         bw.write(String.valueOf((N * N) - (M * M)));
         bw.flush();
         bw.close();
